Adds missing methods with the const_iterator hint.
Added for insert and emplace_hint.
Also minor version updates in the tests and clang-tidy
upgrade doctest

+++ src/test/unit/bench_swap.cpp
@@ -17,7 +17,7 @@ TEST_CASE_TEMPLATE("bench_swap" * doctest::test_suite("nanobench") * doctest::sk
 
     ankerl::nanobench::Bench bench;
     for (size_t i = 0; i < 10; ++i) {
-        for (size_t j = 0; j < 10U * (1U << i); ++j) {
+        for (size_t j = 0; j < static_cast<size_t>(10U) * (1U << i); ++j) {
             a[rng()];
             b[rng()];
         }
